+++ CHANGELOG.md
@@ -10,13 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update group admin API [#518](https://github.com/rokwire/core-building-block/issues/518)
 - Update role permissions admin API [#516](https://github.com/rokwire/core-building-block/issues/516)
 
+### Fixed
+- Update permission should change roles/groups/accounts [#542](https://github.com/rokwire/core-building-block/issues/542)
+
 ## [1.24.2] - 2022-08-08
 ### Added
 - Allow passing nil context to WithContext storage functions [#494](https://github.com/rokwire/core-building-block/issues/494)
 - Account system configs [#558](https://github.com/rokwire/core-building-block/issues/558)
 
 ### Fixed
-- Update permission should change roles/groups/accounts [#542](https://github.com/rokwire/core-building-block/issues/542)
 - Authorization policy comments not working [#506](https://github.com/rokwire/core-building-block/issues/506)
 
 ## [1.24.1] - 2022-07-07